Fixes for pvc stress test
Refactor to move tests for
Create/Delete (CD) and Create/Read/Update/Delete (CRUD)
into a single file.
CD tests are performed prior to CRUD
The number of cycles for CD and CRUD tests can now be
set using environment variables

Update README.md
Remove test runner script

+++ test/e2e/nightly/pvc_stress_fio/pvc_stress_fio_test.go
@@ -3,6 +3,8 @@ package pvc_stress_fio_test
 
 import (
 	"fmt"
+	"os"
+	"strconv"
 	"testing"
 
 	Cmn "e2e-basic/common"
@@ -18,19 +20,35 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-var defTimeoutSecs = "30s"
+var defTimeoutSecs = "60s"
 
-// Create a PVC and verify that
+// Create Delete iterations
+var cdIterations = 100
+
+// Create Read Update Delete iterations
+var crudIterations = 100
+
+// volume name and associated storage class name
+// parameters required by RmPVC
+type volSc struct {
+	volName string
+	scName  string
+}
+
+var podNames []string
+var volNames []volSc
+
+// Create a PVC and verify that (also see and keep in sync with README.md#pvc_stress_fio)
 //	1. The PVC status transitions to bound,
 //	2. The associated PV is created and its status transitions bound
 //	3. The associated MV is created and has a State "healthy"
-//  4. That a test application (fio) can read and write to the volume
+//  4. Optionally that a test application (fio) can read and write to the volume
 // then Delete the PVC and verify that
 //	1. The PVC is deleted
 //	2. The associated PV is deleted
 //  3. The associated MV is deleted
-func testPVC(volName string, scName string) {
-	fmt.Printf("%s, %s\n", volName, scName)
+func testPVC(volName string, scName string, runFio bool) {
+	fmt.Printf("volume: %s, storageClass:%s, run FIO:%v\n", volName, scName, runFio)
 	// PVC create options
 	createOpts := &coreV1.PersistentVolumeClaim{
 		ObjectMeta: metaV1.ObjectMeta{
@@ -47,7 +65,6 @@ func testPVC(volName string, scName string) {
 			},
 		},
 	}
-
 	// Create the PVC.
 	_, createErr := Cmn.CreatePVC(createOpts)
 	Expect(createErr).To(BeNil())
@@ -56,6 +73,10 @@ func testPVC(volName string, scName string) {
 	pvc, getPvcErr := Cmn.GetPVC(volName)
 	Expect(getPvcErr).To(BeNil())
 	Expect(pvc).ToNot(BeNil())
+
+	// For cleanup
+	tmp := volSc{volName, scName}
+	volNames = append(volNames, tmp)
 
 	// Wait for the PVC to be bound.
 	Eventually(func() coreV1.PersistentVolumeClaimPhase {
@@ -107,26 +128,34 @@ func testPVC(volName string, scName string) {
 		"1s",           // polling interval
 	).Should(Equal("healthy"))
 
-	// Create the fio Pod
-	fioPodName := "fio-" + volName
-	pod, err := Cmn.CreateFioPod(fioPodName, volName)
-	Expect(err).ToNot(HaveOccurred())
-	Expect(pod).ToNot(BeNil())
+	if runFio {
+		// Create the fio Pod
+		fioPodName := "fio-" + volName
+		pod, err := Cmn.CreateFioPod(fioPodName, volName)
+		Expect(err).ToNot(HaveOccurred())
+		Expect(pod).ToNot(BeNil())
 
-	// Wait for the fio Pod to transition to running
-	Eventually(func() bool {
-		return Cmn.IsPodRunning(fioPodName)
-	},
-		defTimeoutSecs,
-		"1s",
-	).Should(Equal(true))
+		// For cleanup
+		podNames = append(podNames, fioPodName)
 
-	// Run the fio test
-	Cmn.RunFio(fioPodName, 5)
+		// Wait for the fio Pod to transition to running
+		Eventually(func() bool {
+			return Cmn.IsPodRunning(fioPodName)
+		},
+			defTimeoutSecs,
+			"1s",
+		).Should(Equal(true))
 
-	// Delete the fio pod
-	err = Cmn.DeletePod(fioPodName)
-	Expect(err).ToNot(HaveOccurred())
+		// Run the fio test
+		Cmn.RunFio(fioPodName, 5)
+
+		// Delete the fio pod
+		err = Cmn.DeletePod(fioPodName)
+		Expect(err).ToNot(HaveOccurred())
+
+		// cleanup
+		podNames = podNames[:len(podNames)-1]
+	}
 
 	// Delete the PVC
 	deleteErr := Cmn.DeletePVC(volName)
@@ -136,8 +165,8 @@ func testPVC(volName string, scName string) {
 	Eventually(func() bool {
 		return Cmn.IsPVCDeleted(volName)
 	},
-		defTimeoutSecs, // timeout
-		"1s",           // polling interval
+		"120s", // timeout
+		"1s",   // polling interval
 	).Should(Equal(true))
 
 	// Wait for the PV to be deleted.
@@ -155,12 +184,19 @@ func testPVC(volName string, scName string) {
 		defTimeoutSecs, // timeout
 		"1s",           // polling interval
 	).Should(Equal(true))
+
+	// cleanup
+	volNames = volNames[:len(volNames)-1]
 }
 
-func stressTestPVC() {
-	for ix := 0; ix < 10; ix++ {
-		testPVC(fmt.Sprintf("stress-pvc-nvmf-%d", ix), "mayastor-nvmf")
-		testPVC(fmt.Sprintf("stress-pvc-iscsi-%d", ix), "mayastor-iscsi")
+func stressTestPVC(iters int, runFio bool) {
+	decoration := ""
+	if runFio {
+		decoration = "-io"
+	}
+	for ix := 1; ix <= iters; ix++ {
+		testPVC(fmt.Sprintf("stress-pvc-nvmf%s-%d", decoration, ix), "mayastor-nvmf", runFio)
+		testPVC(fmt.Sprintf("stress-pvc-iscsi%s-%d", decoration, ix), "mayastor-iscsi", runFio)
 	}
 }
 
@@ -169,20 +205,50 @@ func TestPVCStress(t *testing.T) {
 	RunSpecs(t, "PVC Stress Test Suite")
 }
 
-var _ = Describe("Mayastor PVC Stress test with fio", func() {
-	It("should stress test use of PVCs provisioned over iSCSI and NVMe-of", func() {
-		stressTestPVC()
+var _ = Describe("Mayastor PVC Stress test", func() {
+	It("should stress test creation and deletion of PVCs provisioned over iSCSI and NVMe-of", func() {
+		stressTestPVC(cdIterations, false)
+	})
+
+	It("should stress test creation and deletion of PVCs provisioned over iSCSI and NVMe-of", func() {
+		stressTestPVC(crudIterations, true)
 	})
 })
 
 var _ = BeforeSuite(func(done Done) {
-	logf.SetLogger(zap.LoggerTo(GinkgoWriter, true))
+	logf.SetLogger(zap.New(zap.UseDevMode(true), zap.WriteTo(GinkgoWriter)))
 
 	Cmn.SetupTestEnv()
+	tmp := os.Getenv("e2e_pvc_stress_cd_cycles")
+	if len(tmp) != 0 {
+		var err error
+		cdIterations, err = strconv.Atoi(tmp)
+		Expect(err).NotTo(HaveOccurred())
+		logf.Log.Info("Cycle count changed by environment ", "Create/Delete", cdIterations)
+	}
+
+	tmp = os.Getenv("e2e_pvc_stress_crud_cycles")
+	if len(tmp) != 0 {
+		var err error
+		crudIterations, err = strconv.Atoi(tmp)
+		Expect(err).NotTo(HaveOccurred())
+		logf.Log.Info("Cycle count changed by environment", "Create/Read/Update/Delete", crudIterations)
+	}
+	logf.Log.Info("Number of cycles are", "Create/Delete", cdIterations, "Create/Read/Update/Delete", crudIterations)
+
 	close(done)
 }, 60)
 
 var _ = AfterSuite(func() {
+	// Cleanup resources leftover in the event of failure.
+	for _, pod := range podNames {
+		err := Cmn.DeletePod(pod)
+		Expect(err).ToNot(HaveOccurred())
+	}
+	for _, vol := range volNames {
+		Cmn.RmPVC(vol.volName, vol.scName)
+	}
+
 	// NB This only tears down the local structures for talking to the cluster,
 	// not the kubernetes cluster itself.
 	By("tearing down the test environment")
